+++ components/webapp-mgt/org.wso2.carbon.webapp.list.ui/src/main/resources/web/webapp-list/delete_webapps_ajaxprocessor.jsp
@@ -23,6 +23,13 @@
 <%@ page import="org.wso2.carbon.webapp.list.ui.WebappAdminClient" %>
 <%@ page import="java.util.ResourceBundle" %>
 <%
+
+    String httpMethod = request.getMethod().toLowerCase();
+    if (!"post".equals(httpMethod)) {
+        response.sendError(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
+        return;
+    }
+
     String[] webappkeySet = request.getParameterValues("webappKey");
     String pageNumber = request.getParameter("pageNumber");
     String deleteAllWebapps = request.getParameter("deleteAllWebapps");
